accumulators, core: renamed to ObservableAccumulator.
